pyproject.toml
[project]
name = "fund-portfolio-bot"
version = "0.1.0"
description = "Personal fund portfolio management tool"
requires-python = ">=3.10"
dependencies = [
    "akshare>=1.17.85",
    "exchange-calendars>=4.11.3",
    "httpx>=0.27.0",
    "openai>=1.50.0",
    "pandas>=2.0.0",
    "pydantic>=2.9.0",
    "python-dotenv>=1.2.1",
    "rich>=13.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["F", "E", "I"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"scripts/*.py" = ["T201"]
"src/jobs/*.py" = ["T201"]
"src/adapters/notify/*.py" = ["T201"]
"src/adapters/datasources/*.py" = ["T201"]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[dependency-groups]
dev = [
    "pyright>=1.1.407",
    "ruff>=0.14.5",
]

[tool.pyright]
include = ["src"]
exclude = [".venv", "**/__pycache__"]
pythonVersion = "3.10"
typeCheckingMode = "basic"
reportMissingTypeStubs = false
reportUnknownMemberType = false
reportUnknownVariableType = false
reportOptionalMemberAccess = false  # 忽略 None 的方法访问（依赖注入场景）
reportCallIssue = false              # 忽略 max/min 等函数的重载推断问题
reportArgumentType = false           # 忽略参数类型推断问题
reportOperatorIssue = false          # 忽略运算符类型问题
